apps/ml/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
python-dotenv==1.0.1
